+++ flutter/lib/src/event_processor/flutter_enricher_event_processor.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:sentry/sentry.dart';
 
 import '../navigation/sentry_navigator_observer.dart';
@@ -156,7 +157,7 @@ class FlutterEnricherEventProcessor implements EventProcessor {
       // See https://github.com/flutter/flutter/issues/83919
       // 'window_is_visible': _window.viewConfiguration.visible,
       if (renderer != null) 'renderer': renderer,
-      if (_appFlavor != null) 'appFlavor': _appFlavor!,
+      if (appFlavor != null) 'appFlavor': appFlavor!,
     };
   }
 
@@ -270,10 +271,3 @@ class FlutterEnricherEventProcessor implements EventProcessor {
     return null;
   }
 }
-
-/// Copied from https://api.flutter.dev/flutter/services/appFlavor-constant.html
-/// As soon as Flutter 3.16 is the minimal supported version of Sentry, this
-/// can be replaced with the property from the link above.
-const String? _appFlavor = String.fromEnvironment('FLUTTER_APP_FLAVOR') != ''
-    ? String.fromEnvironment('FLUTTER_APP_FLAVOR')
-    : null;